refactor: change refresh mutation to query
